requirements.txt
python-binance
numpy